+++ instrumentation/google.golang.org/grpc/otelgrpc/grpc_stats_handler_test.go
@@ -262,36 +262,10 @@ func checkServerSpans(t *testing.T, sr *tracetest.SpanRecorder, addr string) {
 }
 
 func checkClientMetrics(t *testing.T, reader metric.Reader, addr, stabilityOptIn string) {
-	var rm metricdata.ResourceMetrics
-	require.Eventually(t, func() bool {
-		rm = metricdata.ResourceMetrics{}
-		if err := reader.Collect(t.Context(), &rm); err != nil {
-			return false
-		}
-		if len(rm.ScopeMetrics) == 0 || len(rm.ScopeMetrics[0].Metrics) == 0 {
-			return false
-		}
-		isOld := stabilityOptIn == "rpc/old" || stabilityOptIn == "rpc/dup"
-		isNew := stabilityOptIn == "" || stabilityOptIn == "rpc/dup"
-
-		var expectedCount int
-		if isOld {
-			expectedCount++
-		}
-		if isNew {
-			expectedCount++
-		}
-		if len(rm.ScopeMetrics[0].Metrics) != expectedCount {
-			return false
-		}
-		for _, m := range rm.ScopeMetrics[0].Metrics {
-			data, ok := m.Data.(metricdata.Histogram[float64])
-			if !ok || len(data.DataPoints) != 5 {
-				return false
-			}
-		}
-		return true
-	}, 1*time.Second, 10*time.Millisecond)
+	rm := metricdata.ResourceMetrics{}
+	err := reader.Collect(t.Context(), &rm)
+	assert.NoError(t, err)
+	require.Len(t, rm.ScopeMetrics, 1)
 
 	host, p, err := net.SplitHostPort(addr)
 	require.NoError(t, err)
@@ -409,16 +383,7 @@ func checkServerMetrics(t *testing.T, reader metric.Reader, stabilityOptIn strin
 		if isNew {
 			expectedCount++
 		}
-		if len(rm.ScopeMetrics[0].Metrics) != expectedCount {
-			return false
-		}
-		for _, m := range rm.ScopeMetrics[0].Metrics {
-			data, ok := m.Data.(metricdata.Histogram[float64])
-			if !ok || len(data.DataPoints) != 5 {
-				return false
-			}
-		}
-		return true
+		return len(rm.ScopeMetrics[0].Metrics) == expectedCount
 	}, 1*time.Second, 10*time.Millisecond)
 
 	require.Len(t, rm.ScopeMetrics, 1)