+++ samplers/jaegerremote/sampler_remote_test.go
@@ -301,11 +301,14 @@ func TestRemotelyControlledSampler_ImmediatelyUpdateOnStartup(t *testing.T) {
 		WithSamplingStrategyFetcher(fetcher),
 		withSamplingStrategyParser(parser),
 	)
-	time.Sleep(100 * time.Millisecond) // waiting for s.pollController
-	sampler.Close()                    // stop pollController, avoid date race
-	s, ok := sampler.sampler.(*rateLimitingSampler)
-	assert.True(t, ok)
-	assert.Equal(t, float64(100), s.maxTracesPerSecond)
+	defer sampler.Close()
+
+	assert.Eventually(t, func() bool {
+		sampler.RLock()
+		defer sampler.RUnlock()
+		s, ok := sampler.sampler.(*rateLimitingSampler)
+		return ok && s.maxTracesPerSecond == 100
+	}, 5*time.Second, 10*time.Millisecond)
 }
 
 func TestRemotelyControlledSampler_multiStrategyResponse(t *testing.T) {